fix: split constants into multiple smaller files (#301)
* fix: optimize constants with normalize seperately

* fix: tsc issue

* fix: watchlist tooltip color
