+++ src/components/SwapWidget/LimitOrder/index.tsx
@@ -6,7 +6,8 @@ import React, { useCallback, useContext, useEffect, useState } from 'react';
 import { Divide, RefreshCcw, X } from 'react-feather';
 import { useTranslation } from 'react-i18next';
 import { ThemeContext } from 'styled-components';
-import { NATIVE, SwapTypes } from 'src/constants';
+import { NATIVE } from 'src/constants';
+import { SwapTypes } from 'src/constants/swap';
 import { useChainId, usePangolinWeb3 } from 'src/hooks';
 import { MixPanelEvents, useMixpanel } from 'src/hooks/mixpanel';
 import { useTokenHook } from 'src/hooks/multiChainsHooks';
