+++ src/components/SwapWidget/MarketOrder/index.tsx
@@ -4,8 +4,10 @@ import React, { useCallback, useContext, useEffect, useMemo, useState } from 're
 import { RefreshCcw } from 'react-feather';
 import { useTranslation } from 'react-i18next';
 import { ThemeContext } from 'styled-components';
-import { SwapTypes, TRUSTED_TOKEN_ADDRESSES, ZERO_ADDRESS } from 'src/constants';
+import { ZERO_ADDRESS } from 'src/constants';
+import { TRUSTED_TOKEN_ADDRESSES } from 'src/constants/address';
 import { DEFAULT_TOKEN_LISTS_SELECTED } from 'src/constants/lists';
+import { SwapTypes } from 'src/constants/swap';
 import { useChainId, usePangolinWeb3 } from 'src/hooks';
 import { useCurrency } from 'src/hooks/Tokens';
 import { MixPanelEvents, useMixpanel } from 'src/hooks/mixpanel';
