+++ src/components/SwapWidget/confirmPriceImpactWithoutFee.ts
@@ -1,5 +1,5 @@
 import { Percent } from '@pangolindex/sdk';
-import { ALLOWED_PRICE_IMPACT_HIGH, PRICE_IMPACT_WITHOUT_FEE_CONFIRM_MIN } from '../../constants';
+import { ALLOWED_PRICE_IMPACT_HIGH, PRICE_IMPACT_WITHOUT_FEE_CONFIRM_MIN } from 'src/constants/swap';
 
 /**
  * Given the price impact, get user confirmation.
