+++ src/components/WalletModal/index.tsx
@@ -9,7 +9,8 @@ import { isMobile } from 'react-device-detect';
 import { useTranslation } from 'react-i18next';
 import { Button } from 'src/components/Button';
 import { avalancheCore, bitKeep, gnosisSafe, hashConnect, injected, talisman, xDefi } from 'src/connectors';
-import { AVALANCHE_CHAIN_PARAMS, IS_IN_IFRAME, SUPPORTED_WALLETS, WalletInfo } from 'src/constants';
+import { IS_IN_IFRAME } from 'src/constants';
+import { AVALANCHE_CHAIN_PARAMS, SUPPORTED_WALLETS, WalletInfo } from 'src/constants/wallets';
 import { MixPanelEvents, useMixpanel } from 'src/hooks/mixpanel';
 // import { AppState, useSelector } from 'src/state';
 import { Box, Modal, ToggleButtons } from '../../';
