+++ src/components/WatchList/CoinChart.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import { Line, LineChart, ResponsiveContainer, Tooltip } from 'recharts';
 import { Box, Button, Text } from 'src/components';
-import { TIMEFRAME } from 'src/constants';
+import { TIMEFRAME } from 'src/constants/swap';
 import { CoingeckoWatchListToken, useCoinGeckoTokenPriceChart } from 'src/state/pcoingecko/hooks';
 import { formattedNum, toNiceDateYear } from 'src/utils/charts';
 import { DurationBtns, SelectedCoinInfo } from './styleds';
@@ -63,7 +63,7 @@ const CoinChart: React.FC<Props> = ({ coin }) => {
             labelFormatter={(_label, data) => {
               return toNiceDateYear(data?.[0]?.payload?.timestamp);
             }}
-            labelStyle={{ paddingTop: 4 }}
+            labelStyle={{ paddingTop: 4, color: 'black' }}
             wrapperStyle={{ top: -70, left: -10, zIndex: 9999 }}
           />
         </LineChart>
