+++ src/hooks/useSwapCallback.ts
@@ -5,7 +5,8 @@ import { parseUnits } from '@ethersproject/units';
 import { CAVAX, JSBI, Percent, Router, SwapParameters, Token, Trade, TradeType } from '@pangolindex/sdk';
 import { useMemo } from 'react';
 import { NEAR_EXCHANGE_CONTRACT_ADDRESS } from 'src/connectors';
-import { BIPS_BASE, INITIAL_ALLOWED_SLIPPAGE, ONE_YOCTO_NEAR, ZERO_ADDRESS } from 'src/constants';
+import { INITIAL_ALLOWED_SLIPPAGE, ONE_YOCTO_NEAR, ZERO_ADDRESS } from 'src/constants';
+import { BIPS_BASE } from 'src/constants/swap';
 import { useGetNearPoolId } from 'src/data/Reserves';
 import { useTransactionAdder } from 'src/state/ptransactions/hooks';
 import { calculateGasMargin, getRouterContract, getRouterContractDaaS, isAddress, shortenAddress } from 'src/utils';
