+++ src/index.tsx
@@ -147,6 +147,9 @@ export function PangolinProvider({
 }
 
 export * from './constants';
+export { SUPPORTED_WALLETS } from './constants/wallets';
+export { ROUTER_ADDRESS, MINICHEF_ADDRESS } from './constants/address';
+export { TIMEFRAME, SwapTypes } from './constants/swap';
 export * from './connectors';
 export * from './components';
 export * from './state/papplication/hooks';
