+++ src/state/pcoingecko/hooks.tsx
@@ -3,8 +3,7 @@ import { CHAINS, ChainId, Currency, Token } from '@pangolindex/sdk';
 import axios, { AxiosResponse } from 'axios';
 import { useEffect, useMemo, useState } from 'react';
 import { UseQueryResult, useQueries, useQuery } from 'react-query';
-import { COINGECKO_CURRENCY_ID, COINGEKO_BASE_URL } from 'src/constants';
-import { COINGECKO_TOKENS_MAPPING } from 'src/constants/coingeckoTokens';
+import { COINGECKO_CURRENCY_ID, COINGECKO_TOKENS_MAPPING, COINGEKO_BASE_URL } from 'src/constants/coingecko';
 import { useChainId } from 'src/hooks';
 import { AppState, useDispatch, useSelector } from 'src/state';
 import { addCoingeckoTokens } from './actions';
